backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
defusedxml==0.7.1
dj-database-url==2.3.0
dj-rest-auth==4.0.1
Django==5.2.5
django-allauth==0.57.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==23.5
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
google-cloud==0.34.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
oauthlib==3.3.1
packaging==25.0
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.1
python-memcached==1.62
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
